Move CollectionLink to own file
