+++ backend/corpora/common/entities/geneset.py
@@ -2,7 +2,7 @@ import typing
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.orm import Session
 
-from ..corpora_orm import DbGeneset, DbGenesetDatasetLink, DbCollectionLink
+from ..corpora_orm import DbGeneset, DbGenesetDatasetLink
 from .entity import Entity
 from ..utils.exceptions import CorporaException
 
@@ -99,16 +99,3 @@ class GenesetDatasetLink(Entity):
             raise CorporaException
 
 
-class CollectionLink(Entity):
-    table = DbCollectionLink
-
-    @classmethod
-    def create(cls, session: Session, collection_id: str, **kwargs) -> "CollectionLink":
-        link = DbCollectionLink(collection_id=collection_id, **kwargs)
-        session.add(link)
-        session.commit()
-        return cls(link)
-
-    @classmethod
-    def retrieve_all_links_for_a_collection(cls, session: Session, collection_id: str) -> typing.List["CollectionLink"]:
-        return session.query(cls.table).filter(cls.table.collection_id == collection_id).all()
