+++ tests/unit/backend/fixtures/generate_data_mixin.py
@@ -1,7 +1,8 @@
 from sqlalchemy.orm import Session
 
 from backend.corpora.common.entities import Collection, Dataset
-from backend.corpora.common.entities.geneset import Geneset, CollectionLink
+from backend.corpora.common.entities.geneset import Geneset
+from backend.corpora.common.entities.collection_link import CollectionLink
 from backend.corpora.common.utils.db_session import db_session_manager
 from tests.unit.backend.utils import (
     BogusCollectionParams,
